cli/requirements.txt
boto3~=1.15.1
PyYAML~=5.3.1
kubernetes~=11.0.0
aws_cdk.core~=1.63.0
aws-cdk.aws-ec2~=1.63.0
aws-cdk.aws-iam~=1.63.0
aws-cdk.aws-efs~=1.63.0
aws-cdk.aws-ecr~=1.63.0
aws-cdk.aws-ecs~=1.63.0
aws-cdk.aws-ssm~=1.63.0
aws_cdk.aws_cognito~=1.63.0
click~=7.1.2
halo~=0.0.30
sh~=1.14.0
